refactor: use boost/core/noncopyable.hpp over boost/noncopyable.hpp
boost/noncopyable.hpp is deprecated
