+++ include/boost/multi_index/detail/copy_map.hpp
@@ -17,12 +17,12 @@
 #include <algorithm>
 #include <boost/core/addressof.hpp>
 #include <boost/core/no_exceptions_support.hpp>
+#include <boost/core/noncopyable.hpp>
 #include <boost/move/core.hpp>
 #include <boost/move/utility_core.hpp>
 #include <boost/multi_index/detail/allocator_traits.hpp>
 #include <boost/multi_index/detail/auto_space.hpp>
 #include <boost/multi_index/detail/raw_ptr.hpp>
-#include <boost/noncopyable.hpp>
 #include <functional>
 
 namespace boost{
