+++ include/boost/multi_index/detail/scoped_bilock.hpp
@@ -13,7 +13,7 @@
 #pragma once
 #endif
 
-#include <boost/noncopyable.hpp>
+#include <boost/core/noncopyable.hpp>
 #include <boost/type_traits/aligned_storage.hpp>
 #include <boost/type_traits/alignment_of.hpp>
 #include <functional>
